Split out overridden functions to new file
