+++ pyproject.toml
@@ -35,6 +35,7 @@ pytest-mock = "^3.11.1"
 pytest-cov = "^4.1.0"
 black = "^23.7.0"
 pytest-watcher = "^0.3.4"
+autoflake = "^2.2.0"
 
 [build-system]
 requires = ["poetry-core"]
